# Conflicts:
#	src/main/java/com/furnitureapp/repository/sales/CartRepository.java

+++ src/main/java/com/furnitureapp/utility/Helper.java
@@ -8,7 +8,7 @@ public class Helper {
         //This generates a fully numeric unique code
         int min = 1;
         int max = 999_999;
-        return ThreadLocalRandom.current().nextLong(min, max + 1);
+        return ThreadLocalRandom.current().nextInt(min, max + 1);
 
         //This is a different way of returning unique numbers only
         //return UUID.randomUUID().toString().replaceAll("[^0-9]", "");
